+++ neural_lam/datastore/mdp.py
@@ -397,13 +397,16 @@ class MDPDatastore(BaseRegularGridDatastore):
             ds_unstacked["state"].isel(time=0).isel(state_feature=0)
         )
         da_domain_allzero = xr.zeros_like(da_state_variable)
-        ds_unstacked["boundary_mask"] = da_domain_allzero.isel(
-            x=slice(self._n_boundary_points, -self._n_boundary_points),
-            y=slice(self._n_boundary_points, -self._n_boundary_points),
-        )
-        ds_unstacked["boundary_mask"] = ds_unstacked.boundary_mask.fillna(
-            1
-        ).astype(int)
+        if self._n_boundary_points == 0:
+            ds_unstacked["boundary_mask"] = da_domain_allzero
+        else:
+            ds_unstacked["boundary_mask"] = da_domain_allzero.isel(
+                x=slice(self._n_boundary_points, -self._n_boundary_points),
+                y=slice(self._n_boundary_points, -self._n_boundary_points),
+            )
+            ds_unstacked["boundary_mask"] = ds_unstacked.boundary_mask.fillna(
+                1
+            ).astype(int)
         return self.stack_grid_coords(da_or_ds=ds_unstacked.boundary_mask)
 
     @property